requirements.txt
flask==2.0.1
flask_login==0.5.0
WTForms==2.3.3
flask_wtf==0.15.1
email_validator==1.1.3
python-decouple==3.4
gunicorn==20.1.0
jinja2==3.1.4
flask-restx==0.5.1
Werkzeug==2.0.3
google-cloud-firestore==2.19.0
python-dotenv==1.0.1
Pyrebase4==4.8.0
cryptography==43.0.1
flask-babel== 4.0.0
hubspot-api-client==10.0.0
firebase-admin==6.6.0
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openpyxl
flask_mail